requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.6
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django_coverage_plugin==3.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
greenlet==3.2.3
kombu==5.5.4
packaging==25.0
phonenumbers==9.0.13
pillow==11.3.0
playwright==1.54.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.22
pyee==13.0.0
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
gunicorn
django-redis
django-filter
boto3
django-storages
ffmpeg-python
factory-boy
requests
gunicorn
drf-spectacular
